backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.20
openai==1.12.0
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0